requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
GeoAlchemy2==0.14.2
Shapely==1.8.2
requests==2.31.0
python-dotenv==1.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
SQLAlchemy==2.0.21
alembic==1.2.0
geopy==2.3.0
numpy>=1.26
pyproj==3.6.0